Fix file permission check in configure()
This check fails when there is an existing directory on `dir_path`.
For example:

    $ stat /tmp/ --format %F
    directory
    $ fluentd -c sigdump.conf
    ...
    2022-02-20 15:39:46 +0900 [error]: config error file="sigdump.conf"
    error_class=Fluent::ConfigError error="'dir_path' doesn't have
    enough permission.: /tmp"

This is because `FileUtils.writiable_p()` check if the ruby process
can write a "file" on that path, so it naturally returns false if
the target path is occupied with a directory.

We can avoid this issue by calling this function with an actual file
path (not the base directory path) that we want to use. Fix it.

Signed-off-by: Fujimoto Seiji <fujimoto@ceptord.net>